Refactored random generator to correct package.
