+++ src/main/java/com/sparta/simulator/controller/CentreAvailability.java
@@ -1,6 +1,7 @@
 package com.sparta.simulator.controller;
 
 import com.sparta.simulator.model.Centre;
+
 import java.util.List;
 
 public class CentreAvailability {
@@ -9,9 +10,9 @@ public class CentreAvailability {
 
         int fullCounter = 0;
         for (Centre c : centreList) {
-                if(!c.isFull()){
-                    fullCounter++;
-                }
+            if (!c.isFull()) {
+                fullCounter++;
+            }
         }
         return fullCounter > 0;
     }
@@ -28,7 +29,7 @@ public class CentreAvailability {
                 }
             }
         } else {
-            for (int i = next+1; i < centreList.size(); i++) {
+            for (int i = next + 1; i < centreList.size(); i++) {
                 if (centreList.get(i).getCentreType().equals(centre)) {
                     if (!centreList.get(i).isFull()) {
                         index = i;
@@ -41,10 +42,10 @@ public class CentreAvailability {
     }
 
     // Based on implementation and current requirements centre will most likely never fill up
-    public static int chosenCentreExistsAndSpaceAvailable(String centre ,List<Centre> centreList){
+    public static int chosenCentreExistsAndSpaceAvailable(String centre, List<Centre> centreList) {
 
         int index = -1;
-        if (chosenCentreExist(centre, centreList, 0) > -1 || spaceAvailable(centreList)){
+        if (chosenCentreExist(centre, centreList, 0) > -1 || spaceAvailable(centreList)) {
             return chosenCentreExist(centre, centreList, 0);
         }
         return index;
