+++ src/main/java/com/sparta/simulator/controller/RandomGenerator.java
@@ -1,25 +1,24 @@
-package com.sparta.simulator.model;
+package com.sparta.simulator.controller;
 
 import java.util.Random;
 
 public class RandomGenerator {
 
     public static String generateRandomCentreTwo() {
-        final String[] centreTypes = {"TechCentre","TrainingHub"};
+        final String[] centreTypes = {"TechCentre", "TrainingHub"};
         Random random = new Random();
         return centreTypes[random.nextInt(centreTypes.length)];
     }
 
     public static String generateRandomCentreThree() {
-        final String[] centreTypes = {"TechCentre","TrainingHub","BootCamp"};
+        final String[] centreTypes = {"TechCentre", "TrainingHub", "BootCamp"};
         Random random = new Random();
         return centreTypes[random.nextInt(centreTypes.length)];
     }
 
     public static int getRandomNumOfTrainees() {
         Random rand = new Random();
-        int numOfTraineesToBeAddedEveryMonth = rand.nextInt(51);
-        return numOfTraineesToBeAddedEveryMonth;
+        return rand.nextInt(51);
     }
 
 }
